requirements.txt
torch>=1.9.0
numpy>=1.19.0
Pillow>=8.0.0
opencv-python>=4.5.0
